+++ backend/src/index.ts
@@ -25,6 +25,7 @@ import Scripting from "./game/scripting";
 import SourceBans from "./net/sourcebans";
 import { Recieves } from "$types/messages";
 import CustomRPC from "./net/rpc";
+import { version } from "../../package.json";
 
 init();
 
@@ -49,8 +50,12 @@ async function init() {
     await Promise.all([ Settings.init(), Values.init() ]);
     if(!Settings.get("finishedSetup")) Log.info("Running in setup mode");
 
+    const lastVersion = Values.get("lastVersion");
+    const updated = lastVersion && Bun.semver.order(version, lastVersion) > 0;
+    if(lastVersion !== version) Values.set("lastVersion", version);
+
     try {
-        Server.init();
+        Server.init(updated);
     } catch {
         Log.error("Failed to start server, is tfview already running?");